% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chirps}
\alias{chirps}
\title{CHIRPs gridded rainfall data}
\description{
A subset of UCSB's CHIRPs gridded daily rainfall data over
Zambia, for the period 25 October, 2016 to 21 November, 2016.
}
\examples{
data(chirps)
}
\author{
Funk et al, 2015
}
\keyword{data}
