% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_noaxes.R
\name{plot_noaxes}
\alias{plot_noaxes}
\title{Plot rasters without axes by default}
\usage{
plot_noaxes(x, axes = FALSE, box = FALSE, mar = c(0, 0, 1, 4), ...)
}
\arguments{
\item{x}{An object of class Raster* or Extent}

\item{axes}{Defaults to FALSE, meaning axes with coordinates are not plotted}

\item{box}{Defaults to FALSE, meaning box drawn is not drawn around plot}

\item{mar}{Sets the default margins for a single plot to c(0, 0, 1, 4)}

\item{...}{Graphical parameters. Any argument that can be passed to
image.plot and to plot, such as main='title', ylab='latitude'}
}
\value{
A plotted raster
}
\description{
This is a small variant on raster::plot that provides default
margin parameters and sets axes and box arguments to false
}
\examples{
data(chirps)
plot_noaxes(x = chirps[[1]])
plot_noaxes(x = chirps[[1:3]])
}
