% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppoly_structure.R
\name{sppoly_structure}
\alias{sppoly_structure}
\title{Evaluate the structure of SpatialPolygons*}
\usage{
sppoly_structure(l)
}
\arguments{
\item{l}{A list of SpatialPolygons* to evaluate}
}
\value{
A matrix
}
\description{
Calculates the number of polygons per SpatialPolygons*, and how
many polygons are in each of those polygons
}
\examples{
data(districts)
polygon_structure(l = list(districts))
}
